libimp in projects

+++ minitalk/libimp/ft_atoi.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dlopez-l <dlopez-l@student.42madrid>       +#+  +:+       +#+        */
+/*   By: dlopez-l <dlopez-l@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/16 14:29:25 by dlopez-l          #+#    #+#             */
-/*   Updated: 2024/01/23 12:32:16 by dlopez-l         ###   ########.fr       */
+/*   Updated: 2024/10/28 16:02:34 by dlopez-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,10 +21,10 @@ static int	ft_isspace(char c)
 	return (0);
 }
 
-int	ft_atoi(const char *nptr)
+long	ft_atoi(const char *nptr)
 {
-	int	aux;
-	int	neg;
+	long	aux;
+	int		neg;
 
 	aux = 0;
 	while (ft_isspace(*nptr))
